chore: back to simplified loss function
